# apps/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Godrop API
  description: |
    On-demand delivery and logistics platform for Nigeria.

    **Authentication domains:**
    - **Customer API** — Phone OTP → JWT. Pass token as `Authorization: Bearer <token>`. Payload: `{ userId }`.
    - **Rider App API** — Phone OTP → JWT. Same header. Payload: `{ riderId, type: "rider" }`.
    - **Vendor Admin API** — Email + password → JWT. Same header. Token payload contains `type: "admin"`.
    - **System Admin API** — Email + password → JWT. Same header. Token payload contains `type: "admin"`.

    **Currency:** All monetary amounts are in **Kobo** (₦ × 100). Never Naira floats.
    **Phone:** E.164 format, e.g. `+2348012345678`.

    **Response envelope:**
    - Success: `{ "success": true, ...fields }` — fields vary by endpoint (see each response schema).
    - Error: `{ "success": false, "error": "message" }`.
    - Single-entity detail responses wrap the entity in a `data` key.
    - Paginated list responses return `data` (array) + `meta` (pagination) at the top level.
  version: 1.0.0
  contact:
    name: Godrop Engineering
    email: dev@godrop.ng
servers:
  - url: https://api.godrop.ng/v1
    description: Production
  - url: https://godrop-production.up.railway.app/api/v1
    description: Staging
  - url: http://localhost:4000/api/v1
    description: Local development

# ─── Security ─────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Used by all three auth domains (customer, vendor admin, system admin). Each token encodes a `type` claim."
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: string, example: "+2348012345678" }
        email: { type: string, format: email }
        avatarUrl: { type: string }
        walletBalanceKobo: { type: integer }
...
paths:
...
  /me:
    get:
      tags: [Profile]
      summary: Get profile
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  id: { type: string }
                  firstName: { type: string }
                  lastName: { type: string }
                  phone: { type: string }
                  email: { type: string, format: email }
                  avatarUrl: { type: string }
                  walletBalanceKobo: { type: integer }
    patch:
      tags: [Profile]
      summary: Update profile
      security: [{ BearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName: { type: string }
                lastName: { type: string }
                email: { type: string, format: email }
                avatarUrl: { type: string }
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  user: { $ref: "#/components/schemas/User" }
...
  /orders:
    get:
      tags: [Orders]
      summary: List orders
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [PENDING, ACCEPTED, PREPARING, READY_FOR_PICKUP, PICKED_UP, IN_TRANSIT, DELIVERED, CANCELLED, FAILED] }
        - in: query
          name: type
          schema: { type: string, enum: [FOOD, GROCERY, RETAIL, PHARMACY, PARCEL, TRUCK] }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Paginated orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Order" }
                  meta: { $ref: "#/components/schemas/Pagination" }
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get order details
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Order with items, timeline, and rider
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  order: { $ref: "#/components/schemas/Order" }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        quantity: { type: integer }
                        unitPriceKobo: { type: integer }
                        totalKobo: { type: integer }
                        notes: { type: string }
                  timeline:
                    type: array
                    items:
                      type: object
                      properties:
                        status: { type: string }
                        createdAt: { type: string, format: date-time }
                  rider:
                    nullable: true
                    type: object
                    properties:
                      id: { type: string }
                      firstName: { type: string }
                      lastName: { type: string }
                      phone: { type: string }
                      lat: { type: number }
                      lng: { type: number }
        "404":
          description: Order not found
  /orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel an order
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  order: { $ref: "#/components/schemas/Order" }
        "404":
          description: Order not found
        "422":
          description: Order cannot be cancelled (already out for delivery or delivered)
...
  /me/notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: isRead
          schema: { type: boolean }
      responses:
        "200":
          description: Notifications with unread count
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Notification" }
                  meta: { $ref: "#/components/schemas/Pagination" }
                  unreadCount: { type: integer, description: "Total unread (regardless of current filter)" }
...
  /admin/login:
    post:
      tags: [Admin — Auth]
      summary: Admin login (system admin and vendor admin)
      description: |
        Unified login endpoint for all admin types. Accepts email + password
        and returns a JWT token valid for both system admins (8h) and vendor
        admins (7d). The `admin.type` field in the response indicates whether
        the caller is a `SYSTEM` or `VENDOR` admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: JWT token and admin profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  token: { type: string }
                  admin:
                    oneOf:
                      - $ref: "#/components/schemas/SystemAdminProfile"
                      - $ref: "#/components/schemas/VendorAdminProfile"
                    discriminator:
                      propertyName: type
                      mapping:
                        SYSTEM: "#/components/schemas/SystemAdminProfile"
                        VENDOR: "#/components/schemas/VendorAdminProfile"
        "401":
          description: Invalid credentials
  /admin/me:
    get:
      tags: [System Admin — Auth]
      summary: Get own system admin profile (ADMIN+)
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: System admin profile (password excluded)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: "#/components/schemas/SystemAdminProfile" }
  /admin/me/change-password:
    post:
      tags: [System Admin — Auth]
      summary: Change own password
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
        "400":
          description: Current password is incorrect
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /admin/me/profile:
    patch:
      tags: [System Admin — Auth]
      summary: Update own profile (firstName, lastName, email)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName: { type: string, minLength: 1, maxLength: 50 }
                lastName: { type: string, minLength: 1, maxLength: 50 }
                email: { type: string, format: email }
      responses:
        "200":
          description: Updated system admin profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: "#/components/schemas/SystemAdminProfile" }
        "409":
          description: Email already in use
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /admin/me/settings:
    get:
      tags: [System Admin — Auth]
      summary: Get own notification/report settings
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Admin settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: "#/components/schemas/SystemAdminSettings" }
    patch:
      tags: [System Admin — Auth]
      summary: Update own notification/report settings
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                emailNotifications: { type: boolean }
                weeklyReport: { type: boolean }
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: "#/components/schemas/SystemAdminSettings" }

  # ─ Admin Management (SUPER_ADMIN only) ───────────────────────
  /admin/admins:
    get:
      tags: [System Admin — Admins]
      summary: List system admins (SUPER_ADMIN only)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Paginated list of system admins
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/SystemAdminProfile" }
                  total: { type: integer }
                  page: { type: integer }
                  limit: { type: integer }
        "403":
          description: Forbidden — requires SUPER_ADMIN role
    post:
      tags: [System Admin — Admins]
      summary: Create system admin (SUPER_ADMIN only)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, firstName, lastName, password]
              properties:
                email: { type: string, format: email }
                firstName: { type: string, maxLength: 50 }
                lastName: { type: string, maxLength: 50 }
                password: { type: string, minLength: 8 }
                role:
                  type: string
                  enum: [SUPER_ADMIN, ADMIN]
                  default: ADMIN
      responses:
        "201":
          description: Admin created (password excluded)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: "#/components/schemas/SystemAdminProfile" }
        "403":
          description: Forbidden — requires SUPER_ADMIN role
        "409":
          description: Email already registered
  /admin/admins/{id}:
    patch:
      tags: [System Admin — Admins]
      summary: Update system admin (SUPER_ADMIN only)
      description: Can update name, active status, and role.
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName: { type: string, maxLength: 50 }
                lastName: { type: string, maxLength: 50 }
                isActive: { type: boolean }
                role: { type: string, enum: [SUPER_ADMIN, ADMIN] }
      responses:
        "200":
          description: Updated admin profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: "#/components/schemas/SystemAdminProfile" }
        "403":
          description: Forbidden — requires SUPER_ADMIN role

  # ─ Vendor Management (ADMIN+) ─────────────────────────────────
  /admin/vendors:
    get:
      tags: [System Admin — Vendors]
      summary: List all vendors (ADMIN+)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [PENDING, APPROVED, REJECTED, SUSPENDED] }
        - in: query
          name: type
          schema: { type: string, enum: [RESTAURANT, GROCERY, RETAIL, PHARMACY] }
        - in: query
          name: search
          description: Searches by vendor name (case-insensitive)
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Paginated vendor list with owner info and counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/VendorSummary" }
                  total: { type: integer }
                  page: { type: integer }
                  limit: { type: integer }
  /admin/vendors/{id}:
    get:
      tags: [System Admin — Vendors]
      summary: Get vendor detail (ADMIN+)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Full vendor with all team members and counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: "#/components/schemas/VendorSummary" }
  /admin/vendors/{id}/approve:
    patch:
      tags: [System Admin — Vendors]
      summary: Approve vendor (ADMIN+)
      description: |
        Sets status to APPROVED and `isActive = true`.
        Sends an approval email to the vendor owner.
        This action is audit-logged.
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Vendor approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: "#/components/schemas/VendorFull" }
        "409":
          description: Vendor is already approved
  /admin/vendors/{id}/reject:
    patch:
      tags: [System Admin — Vendors]
      summary: Reject vendor (ADMIN+)
      description: |
        Sets status to REJECTED and `isActive = false`. Stores the rejection reason.
        Sends a rejection email to the vendor owner.
        This action is audit-logged.
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  example: "Incomplete documentation submitted"
      responses:
        "200":
          description: Vendor rejected
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: "#/components/schemas/VendorFull" }
        "409":
          description: Vendor is already rejected
  /admin/vendors/{id}/suspend:
    patch:
      tags: [System Admin — Vendors]
      summary: Suspend vendor (ADMIN+)
      description: |
        Sets status to SUSPENDED and `isActive = false`.
        The vendor immediately disappears from customer listings.
        This action is audit-logged.
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          description: Vendor suspended
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: "#/components/schemas/VendorFull" }
        "409":
          description: Vendor is already suspended
  /admin/vendors/{id}/reinstate:
    patch:
      tags: [System Admin — Vendors]
      summary: Reinstate vendor (ADMIN+)
      description: |
        Sets status back to APPROVED and `isActive = true`.
        Use this to un-suspend a previously suspended vendor.
        This action is audit-logged.
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Vendor reinstated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: "#/components/schemas/VendorFull" }

  # ─ Audit Logs (ADMIN+) ────────────────────────────────────────
  /admin/audit-logs:
    get:
      tags: [System Admin — Audit Logs]
      summary: Query audit logs (ADMIN+)
      description: |
        Returns a paginated, filterable audit trail of all actions taken by
        vendor admins and system admins. Sorted by most recent first.
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: vendorId
          description: Filter by vendor
          schema: { type: string }
        - in: query
          name: adminId
          description: Filter by actor (vendor admin or system admin) ID
          schema: { type: string }
        - in: query
          name: action
          description: Partial match on action string (case-insensitive), e.g. "APPROVE"
          schema: { type: string }
        - in: query
          name: entity
          description: Partial match on entity name, e.g. "Vendor"
          schema: { type: string }
        - in: query
          name: from
          description: ISO 8601 datetime — log entries on or after this time
          schema: { type: string, format: date-time }
        - in: query
          name: to
          description: ISO 8601 datetime — log entries on or before this time
          schema: { type: string, format: date-time }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: Paginated audit log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/AuditLog" }
                  total: { type: integer }
                  page: { type: integer }
                  limit: { type: integer }

  # ─ Orders (ADMIN+) ───────────────────────────────────────────
  /admin/orders:
    get:
      tags: [System Admin — Orders]
      summary: List all platform orders (ADMIN+)
      description: |
        Returns a paginated, filterable list of all orders across all vendors.
        Sorted by most recent first.
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, ACCEPTED, PREPARING, READY_FOR_PICKUP, PICKED_UP, IN_TRANSIT, DELIVERED, CANCELLED, FAILED]
        - in: query
          name: type
          schema:
            type: string
            enum: [FOOD, GROCERY, RETAIL, PHARMACY, PARCEL, TRUCK]
        - in: query
          name: vendorId
          schema: { type: string }
        - in: query
          name: search
          description: Search by tracking code, customer name, or phone
          schema: { type: string }
        - in: query
          name: from
          description: "ISO date (YYYY-MM-DD) — orders created on or after this date"
          schema: { type: string, format: date }
        - in: query
          name: to
          description: "ISO date (YYYY-MM-DD) — orders created on or before this date"
          schema: { type: string, format: date }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Paginated platform orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/AdminOrder" }
                  total: { type: integer }
                  page: { type: integer }
                  limit: { type: integer }
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — requires ADMIN role or higher
  /admin/orders/{id}:
    get:
      tags: [System Admin — Orders]
      summary: Get order detail (ADMIN+)
      description: |
        Returns the full order record including customer, vendor, rider, line items,
        status timeline, and — for TRUCK orders — apartment type, loader count, and stops.
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Full order detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: "#/components/schemas/AdminOrder" }
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — requires ADMIN role or higher
        "404":
          description: Order not found
...
  /admin/customers/{id}:
    get:
      tags: [System Admin — Customers]
      summary: Get customer detail (ADMIN+)
      description: Returns full profile, wallet balance, and order/address counts for a single customer.
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: "#/components/schemas/AdminCustomerDetail" }
        "404":
          description: Customer not found
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — requires ADMIN role or higher
...